Pulled out the HTML generation into it's own file
